Add technical revision to allow author to edit metatada

Closes specs#517

See merge request wjs/wjs-profile-project!291

+++ wjs/plugins/wjs_review/tests/test_logic.py
@@ -10,7 +10,6 @@ from django.urls import reverse
 from django.utils.timezone import now
 from faker import Faker
 from review import models as review_models
-from review.const import EditorialDecisions
 from review.models import ReviewAssignment, ReviewForm
 from submission import models as submission_models
 from submission.models import Keyword
@@ -956,6 +955,7 @@ def test_submit_review_messages(
         (ArticleWorkflow.Decisions.NOT_SUITABLE, ArticleWorkflow.ReviewStates.NOT_SUITABLE),
         (ArticleWorkflow.Decisions.MINOR_REVISION, ArticleWorkflow.ReviewStates.TO_BE_REVISED),
         (ArticleWorkflow.Decisions.MAJOR_REVISION, ArticleWorkflow.ReviewStates.TO_BE_REVISED),
+        (ArticleWorkflow.Decisions.TECHNICAL_REVISION, ArticleWorkflow.ReviewStates.TO_BE_REVISED),
     ),
 )
 @pytest.mark.django_db
@@ -972,7 +972,7 @@ def test_handle_editor_decision(
     If the editor makes a decision, article.stage is set to the next workflow stage if decision is final
     and articleworkflow.state is updated according to the decision.
     """
-    section_editor = assigned_article.editorassignment_set.first().editor
+    editor_user = assigned_article.editorassignment_set.first().editor
     fake_request.user = jcom_user
     review_2 = _create_review_assignment(
         fake_request=fake_request,
@@ -986,7 +986,7 @@ def test_handle_editor_decision(
     assert assigned_article.reviewassignment_set.filter(date_declined__isnull=False).count() == 0
     assert assigned_article.reviewassignment_set.filter(is_complete=True).count() == 1
 
-    fake_request.user = section_editor
+    fake_request.user = editor_user
     form_data = {
         "decision": decision,
         "decision_editor_report": "random message",
@@ -1005,14 +1005,15 @@ def test_handle_editor_decision(
     handle = HandleDecision(
         workflow=assigned_article.articleworkflow,
         form_data=form_data,
-        user=section_editor,
+        user=editor_user,
         request=fake_request,
     )
     handle.run()
     assigned_article.refresh_from_db()
 
-    # We had an incomplete review assignement; when the editor makes a decision, these are withdrawn and the reviewer
-    # informed. Let's drop the Message and the email for the reviewers to ease testing.
+    # We had an incomplete review assignment; when the editor makes a decision (except for technical revision),
+    # these are withdrawn and the reviewer informed.
+    # Test this case for all the states at once to ease detailed testing.
     message_template_context = {
         "article": assigned_article,
         "request": fake_request,
@@ -1037,23 +1038,26 @@ def test_handle_editor_decision(
         context=message_template_context,
         template_is_setting=True,
     )
-    assert Message.objects.count() == 2
-    withdrawn_review_message = Message.objects.order_by("created").first()
-    assert withdrawn_review_message.subject == review_withdraw_message_subject
-    assert withdrawn_review_message.body == review_withdraw_message_body
-    assert withdrawn_review_message.message_type == Message.MessageTypes.VERBOSE
-    assert len(mail.outbox) == 2
-    withdrawn_review_mail = mail.outbox[0]
-    assert review_withdraw_message_subject in withdrawn_review_mail.subject
-    assert review_withdraw_message_body in withdrawn_review_mail.body
+    if decision == ArticleWorkflow.Decisions.TECHNICAL_REVISION:
+        assert Message.objects.count() == 1
+    elif decision not in (ArticleWorkflow.Decisions.TECHNICAL_REVISION,):
+        assert Message.objects.count() == 2
+        withdrawn_review_message = Message.objects.order_by("created").first()
+        assert withdrawn_review_message.subject == review_withdraw_message_subject
+        assert withdrawn_review_message.body == review_withdraw_message_body
+        assert withdrawn_review_message.message_type == Message.MessageTypes.VERBOSE
+        assert len(mail.outbox) == 2
+        withdrawn_review_mail = mail.outbox[0]
+        assert review_withdraw_message_subject in withdrawn_review_mail.subject
+        assert review_withdraw_message_body in withdrawn_review_mail.body
 
-    Message.objects.get(subject=review_withdraw_message_subject).delete()
-    for message in mail.outbox:
-        if review_withdraw_message_subject in message.subject:
-            mail.outbox.remove(message)
-            break
+        Message.objects.get(subject=review_withdraw_message_subject).delete()
+        for message in mail.outbox:
+            if review_withdraw_message_subject in message.subject:
+                mail.outbox.remove(message)
+                break
 
-    if final_state == ArticleWorkflow.ReviewStates.TO_BE_REVISED:
+    if decision in (ArticleWorkflow.Decisions.MAJOR_REVISION, ArticleWorkflow.Decisions.MINOR_REVISION):
         # article is kept the as ON_WORKFLOW_ELEMENT_COMPLETE event is not triggered
         assert assigned_article.stage == submission_models.STAGE_UNDER_REVISION
         assert assigned_article.articleworkflow.state == final_state
@@ -1063,27 +1067,20 @@ def test_handle_editor_decision(
         )
         assert revision.editor_note == "random message"
         assert revision.date_due == form_data["date_due"]
-        assert revision.type == (
-            EditorialDecisions.MINOR_REVISIONS.value
-            if form_data["decision"] == ArticleWorkflow.Decisions.MINOR_REVISION
-            else EditorialDecisions.MAJOR_REVISIONS.value
-        )
+        assert revision.type == form_data["decision"]
         # Prepare subjects and bodies
-        message_template_context.update(
-            {
-                "revision": revision,
-                "major_revision": revision.type
-                in (
-                    ArticleWorkflow.Decisions.MAJOR_REVISION.value,
-                    EditorialDecisions.MAJOR_REVISIONS.value,
-                ),
-                "minor_revision": revision.type
-                in (
-                    ArticleWorkflow.Decisions.MINOR_REVISION.value,
-                    EditorialDecisions.MINOR_REVISIONS.value,
-                ),
-            },
-        )
+        message_template_context = {
+            "article": assigned_article,
+            "request": fake_request,
+            "revision": revision,
+            "major_revision": revision.type == ArticleWorkflow.Decisions.MAJOR_REVISION,
+            "minor_revision": revision.type == ArticleWorkflow.Decisions.MINOR_REVISION,
+            "tech_revision": revision.type == ArticleWorkflow.Decisions.TECHNICAL_REVISION,
+            "decision": form_data["decision"],
+            "user_message_content": form_data["decision_editor_report"],
+            "withdraw_notice": form_data["withdraw_notice"],
+            "skip": False,
+        }
         revision_request_message_subject = render_template_from_setting(
             setting_group_name="wjs_review",
             setting_name="review_decision_revision_request_subject",
@@ -1100,18 +1097,18 @@ def test_handle_editor_decision(
             context=message_template_context,
             template_is_setting=True,
         )
-        # Check the messages
+        # Check the messages - withdraw message testsed above
         assert Message.objects.count() == 1
-        revision_request_message = Message.objects.get()
+        revision_request_message = Message.objects.order_by("created").last()
         assert revision_request_message.subject == revision_request_message_subject
         assert revision_request_message.body == revision_request_message_body
         assert revision_request_message.message_type == Message.MessageTypes.VERBOSE
-        # Check the emails
+        # Check the emails - withdraw message testsed above
         assert len(mail.outbox) == 1
         revision_request_mail = mail.outbox[0]
         assert revision_request_message_subject in revision_request_mail.subject
         assert revision_request_message_body in revision_request_mail.body
-    elif final_state == ArticleWorkflow.ReviewStates.ACCEPTED:
+    elif decision == ArticleWorkflow.Decisions.ACCEPT:
         # article is moved to the next stage by ON_WORKFLOW_ELEMENT_COMPLETE event triggered by HandleDecision
         next_stage = get_next_workflow(assigned_article.journal)
         assert assigned_article.stage == next_stage.stage
@@ -1141,7 +1138,7 @@ def test_handle_editor_decision(
         # Check the message
         assert Message.objects.count() == 1
         accept_message = Message.objects.get()
-        assert accept_message.actor == section_editor
+        assert accept_message.actor == editor_user
         assert list(accept_message.recipients.all()) == [assigned_article.correspondence_author]
         assert accept_message.subject == accept_message_subject
         assert accept_message.body == accept_message_body
@@ -1151,7 +1148,7 @@ def test_handle_editor_decision(
         accept_mail = mail.outbox[0]
         assert accept_message_subject in accept_mail.subject
         assert accept_message_body in accept_mail.body
-    elif final_state == ArticleWorkflow.ReviewStates.NOT_SUITABLE:
+    elif decision == ArticleWorkflow.Decisions.NOT_SUITABLE:
         assert assigned_article.stage == submission_models.STAGE_REJECTED
         assert assigned_article.articleworkflow.state == final_state
         # Prepare subject and body
@@ -1179,7 +1176,7 @@ def test_handle_editor_decision(
         )
         # Check the message
         assert Message.objects.count() == 1
-        assert not_suitable_message.actor == section_editor
+        assert not_suitable_message.actor == editor_user
         assert list(not_suitable_message.recipients.all()) == [assigned_article.correspondence_author]
         assert not_suitable_message.subject == not_suitable_message_subject
         assert not_suitable_message.body == not_suitable_message_body
@@ -1188,7 +1185,7 @@ def test_handle_editor_decision(
         not_suitable_mail = mail.outbox[0]
         assert not_suitable_message_subject in not_suitable_mail.subject
         assert not_suitable_message_body in not_suitable_mail.body
-    elif final_state == ArticleWorkflow.ReviewStates.REJECTED:
+    elif decision == ArticleWorkflow.Decisions.REJECT:
         assert assigned_article.stage == submission_models.STAGE_REJECTED
         assert assigned_article.articleworkflow.state == final_state
         # Prepare subject and body
@@ -1216,7 +1213,7 @@ def test_handle_editor_decision(
         # Check the message
         assert Message.objects.count() == 1
         reject_message = Message.objects.get()
-        assert reject_message.actor == section_editor
+        assert reject_message.actor == editor_user
         assert list(reject_message.recipients.all()) == [assigned_article.correspondence_author]
         assert reject_message.subject == reject_message_subject
         assert reject_message.body == reject_message_body
@@ -1226,13 +1223,59 @@ def test_handle_editor_decision(
         reject_mail = mail.outbox[0]
         assert reject_message_subject in reject_mail.subject
         assert reject_message_body in reject_mail.body
+    elif decision == ArticleWorkflow.Decisions.TECHNICAL_REVISION:
+        assert assigned_article.stage == submission_models.STAGE_UNDER_REVIEW
+        assert assigned_article.articleworkflow.state == final_state
+        # Prepare subject and body
+        technical_revision_message_subject = get_setting(
+            setting_group_name="wjs_review",
+            setting_name="technical_revision_subject",
+            journal=assigned_article.journal,
+        ).processed_value
+        technical_revision_message_body = render_template_from_setting(
+            setting_group_name="wjs_review",
+            setting_name="technical_revision_body",
+            journal=assigned_article.journal,
+            request=fake_request,
+            context={
+                "article": assigned_article,
+                "request": fake_request,
+                "revision": None,
+                "decision": form_data["decision"],
+                "user_message_content": form_data["decision_editor_report"],
+                "withdraw_notice": form_data["withdraw_notice"],
+                "skip": False,
+            },
+            template_is_setting=True,
+        )
+        # Check the message
+        assert Message.objects.count() == 1
+        technical_revision_message = Message.objects.get()
+        assert technical_revision_message.actor == editor_user
+        assert list(technical_revision_message.recipients.all()) == [assigned_article.correspondence_author]
+        assert technical_revision_message.subject == technical_revision_message_subject
+        assert technical_revision_message.body == technical_revision_message_body
+        assert technical_revision_message.message_type == Message.MessageTypes.VERBOSE
+        # Check that one email is sent by us (and not by Janeway)
+        assert len(mail.outbox) == 1
+        reject_mail = mail.outbox[0]
+        assert technical_revision_message_subject in reject_mail.subject
+        assert technical_revision_message_body in reject_mail.body
 
-    # All review assignments are marked as complete; the one that was pending when the editor decision was take is
-    # marked as "withdrawn".
-    assert assigned_article.reviewassignment_set.filter(date_accepted__isnull=True).count() == 1
-    assert assigned_article.reviewassignment_set.filter(date_declined__isnull=False).count() == 0
-    assert assigned_article.reviewassignment_set.filter(is_complete=True).count() == 2
-    assert assigned_article.reviewassignment_set.filter(decision="withdrawn").count() == 1
+    if decision == ArticleWorkflow.Decisions.TECHNICAL_REVISION:
+        # All review assignments are marked as complete; the one that was pending when the editor decision was take is
+        # marked as "withdrawn".
+        assert assigned_article.reviewassignment_set.filter(date_accepted__isnull=True).count() == 1
+        assert assigned_article.reviewassignment_set.filter(date_declined__isnull=False).count() == 0
+        assert assigned_article.reviewassignment_set.filter(is_complete=True).count() == 1
+        assert assigned_article.reviewassignment_set.filter(decision="withdrawn").count() == 0
+    else:
+        # All review assignments are marked as complete; the one that was pending when the editor decision was take is
+        # marked as "withdrawn".
+        assert assigned_article.reviewassignment_set.filter(date_accepted__isnull=True).count() == 1
+        assert assigned_article.reviewassignment_set.filter(date_declined__isnull=False).count() == 0
+        assert assigned_article.reviewassignment_set.filter(is_complete=True).count() == 2
+        assert assigned_article.reviewassignment_set.filter(decision="withdrawn").count() == 1
 
     editor_decision = EditorDecision.objects.get(
         workflow=assigned_article.articleworkflow,
